Moved compute transaction to separate function
