+++ stocktracker/__init__.py
@@ -1,8 +1,7 @@
 """StockTracker Main.py"""
 # pylint: disable=line-too-long
 # pylint: disable=logging-fstring-interpolation
-# pylint: disable=too-many-locals
-# pylint: disable=duplicate-code
+
 
 # Import modules
 import logging
@@ -13,201 +12,6 @@ import pandas
 from shared_code import get_config
 
 # modules
-def compute_transactions(transactions):
-    """Compute transactions"""
-    logging.info("Computing transactions")
-    transactions = sorted(
-        transactions["transactions"], key=lambda k: k["transaction_date"]
-    )
-    stocks_held = get_transactions_by_day(transactions)
-    stocks_held = calculate_sells_and_buys(stocks_held)
-    stocks_held = merge_sells_and_buys(stocks_held)
-    return stocks_held
-
-
-def get_transactions_by_day(transactions):
-    """Get transactions by day"""
-    logging.info("Getting transactions by day")
-    # initialize variables
-    stocks_held = {}
-
-    # grab dates
-    end_date = date.today()
-    start_date = transactions[0]["transaction_date"]
-    daterange = pandas.date_range(start_date, end_date)
-
-    # loop through dates
-    for single_date in daterange:
-        logging.debug(f"Getting transactions for {single_date}")
-        single_date = single_date.strftime("%Y-%m-%d")
-        filterd_stocks_held = [
-            d for d in transactions if d["transaction_date"] <= single_date
-        ]
-
-        # create object
-        temp_list = []
-        for filterd_stock_held in filterd_stocks_held:
-            temp_object = {
-                "symbol": filterd_stock_held["symbol"],
-                "cost": filterd_stock_held["cost"],
-                "quantity": filterd_stock_held["quantity"],
-                "transaction_type": filterd_stock_held["transaction_type"],
-                "transaction_cost": filterd_stock_held["transaction_cost"],
-                "currency": filterd_stock_held["currency"],
-            }
-            temp_list.append(temp_object)
-
-        stocks_held.update({single_date: temp_list})
-    # return dictionary
-    stocks_held = {"stocks_held": stocks_held}
-    return stocks_held
-
-
-def calculate_sells_and_buys(stocks_held):
-    """Merge sells and buys together"""
-    logging.info("Calculating sells and buys")
-    # initialize variables
-    computed_date_stocks_held = {}
-
-    # Loop through dates
-    for single_date, date_stocks_held in stocks_held["stocks_held"].items():
-        logging.debug(f"Calculating sells and buys for {single_date}")
-        # initialize variables
-        symbols_buys = []
-        symbols_sells = []
-        temp_list = []
-
-        # get buys
-        date_stocks_held_buys = [
-            d for d in date_stocks_held if d["transaction_type"] == "Buy"
-        ]
-
-        # get symbols
-        for temp_loop in date_stocks_held_buys:
-            symbols_buys.append(temp_loop["symbol"])
-            symbols_buys = list(dict.fromkeys(symbols_buys))
-
-        # create computed object
-        for symbol_buys in symbols_buys:
-            date_stock_held_buys = [
-                d for d in date_stocks_held_buys if d["symbol"] == symbol_buys
-            ]
-            if not date_stock_held_buys:
-                continue
-            temp_object = {
-                "symbol": symbol_buys,
-                "average_cost": sum([d["cost"] for d in date_stock_held_buys])
-                / sum([d["quantity"] for d in date_stock_held_buys]),
-                "quantity": sum([d["quantity"] for d in date_stock_held_buys]),
-                "transaction_type": "Buy",
-                "transaction_cost": sum(
-                    [d["transaction_cost"] for d in date_stock_held_buys]
-                ),
-                "currency": date_stock_held_buys[0]["currency"],
-            }
-            temp_list.append(temp_object)
-
-        # Get sells
-        date_stocks_held_sells = [
-            d for d in date_stocks_held if d["transaction_type"] == "Sell"
-        ]
-
-        # get symbols
-        for temp_loop in date_stocks_held_sells:
-            symbols_sells.append(temp_loop["symbol"])
-            symbols_sells = list(dict.fromkeys(symbols_sells))
-
-        # create computed object
-        for symbol_sells in symbols_sells:
-            date_stock_held_sells = [
-                d for d in date_stocks_held_sells if d["symbol"] == symbol_sells
-            ]
-            if not date_stock_held_sells:
-                continue
-            temp_object = {
-                "symbol": symbol_sells,
-                "average_cost": sum([d["cost"] for d in date_stock_held_sells])
-                / sum([d["quantity"] for d in date_stock_held_sells]),
-                "quantity": sum([d["quantity"] for d in date_stock_held_sells]),
-                "transaction_type": "Sell",
-                "transaction_cost": sum(
-                    [d["transaction_cost"] for d in date_stock_held_sells]
-                ),
-                "currency": date_stock_held_sells[0]["currency"],
-            }
-            temp_list.append(temp_object)
-        computed_date_stocks_held.update({single_date: temp_list})
-
-    # return dictionary
-    computed_date_stocks_held = {"stocks_held": computed_date_stocks_held}
-    return computed_date_stocks_held
-
-
-def merge_sells_and_buys(stocks_held):
-    """Loop through buys and sells and merge them together"""
-    # pylint: disable=too-many-locals
-
-    logging.info("Merging sells and buys")
-
-    # initialize variables
-    merged_stocks_held = {}
-    uid = 0
-    # loop through dates
-    for single_date, date_stocks_held in stocks_held["stocks_held"].items():
-        logging.debug(f"Merging sells and buys for {single_date}")
-
-        # initialize variables
-        symbols = []
-        temp_list = []
-
-        # get symbols
-        for temp_loop in date_stocks_held:
-            symbols.append(temp_loop["symbol"])
-            symbols = list(dict.fromkeys(symbols))
-
-        # loop through symbols
-        for symbol in symbols:
-            single_stock_list = [d for d in date_stocks_held if d["symbol"] == symbol]
-
-            if (
-                len(single_stock_list) == 1
-                and single_stock_list[0]["transaction_type"] == "Buy"
-            ):
-                temp_object = {
-                    "uid": uid,
-                    "symbol": symbol,
-                    "average_cost": single_stock_list[0]["average_cost"],
-                    "total_cost": single_stock_list[0]["average_cost"]
-                    * single_stock_list[0]["quantity"],
-                    "quantity": single_stock_list[0]["quantity"],
-                    "transaction_cost": single_stock_list[0]["transaction_cost"],
-                    "currency": single_stock_list[0]["currency"],
-                }
-                temp_list.append(temp_object)
-            elif len(single_stock_list) == 2:
-                single_stock_list = sorted(
-                    single_stock_list, key=lambda k: k["transaction_type"]
-                )
-                temp_object = {
-                    "uid": uid,
-                    "symbol": symbol,
-                    "average_cost": single_stock_list[0]["average_cost"],
-                    "total_cost": single_stock_list[0]["average_cost"]
-                    * single_stock_list[0]["quantity"],
-                    "quantity": single_stock_list[0]["quantity"]
-                    - single_stock_list[1]["quantity"],
-                    "transaction_cost": single_stock_list[0]["transaction_cost"]
-                    + single_stock_list[1]["transaction_cost"],
-                    "currency": single_stock_list[0]["currency"],
-                }
-                if temp_object["quantity"] > 0:
-                    temp_list.append(temp_object)
-            uid += 1
-        merged_stocks_held.update({single_date: temp_list})
-    merged_stocks_held = {"stocks_held": merged_stocks_held}
-    return merged_stocks_held
-
-
 def calculate_totals(stocks_held):
     """Calculate totals"""
     logging.info("Calculating totals")
@@ -538,10 +342,9 @@ def main(name: str) -> str:
     transactions = json.loads(name[0])
     stock_data = json.loads(name[1])
     forex_data = json.loads(name[2])
-
+    stock_held = json.loads(name[3])
 
     # build data
-    stock_held = compute_transactions(transactions)
     invested = get_invested_data(transactions)
     data = add_stock_data_to_stocks_held(stock_held, stock_data, forex_data)
     data = calculate_totals(data)
